refactor(dashboard): extract reusable EmbedSnippet, AuthCard, FormField
- Pull the install-snippet UI out of EmbedCard into a route-agnostic EmbedSnippet (one source for the embed markup; onboarding reuses it)
- Add AuthCard + FormField shells shared by the auth screens
