+++ apps/dashboard/src/components/auth-card.tsx
@@ -1,0 +1,30 @@
+import {
+	Card,
+	CardContent,
+	CardDescription,
+	CardHeader,
+	CardTitle,
+} from "@/components/ui/card";
+
+/** Centered card shell shared by the sign-in and sign-up screens. */
+export function AuthCard({
+	title,
+	description,
+	children,
+}: {
+	title: string;
+	description: string;
+	children: React.ReactNode;
+}) {
+	return (
+		<main className="flex min-h-screen items-center justify-center p-6">
+			<Card className="w-full max-w-sm">
+				<CardHeader>
+					<CardTitle className="text-xl">{title}</CardTitle>
+					<CardDescription>{description}</CardDescription>
+				</CardHeader>
+				<CardContent>{children}</CardContent>
+			</Card>
+		</main>
+	);
+}
